[CB-329] add event page for coffee event and change name of the event page to event list
