+++ lib/page/home.dart
@@ -7,7 +7,7 @@ import 'package:chocobread/page/alertnoservice.dart';
 import 'package:chocobread/page/blockuser.dart';
 import 'package:chocobread/page/colordeterminants/colorstatus.dart';
 import 'package:chocobread/page/detail.dart';
-import 'package:chocobread/page/event/event_page.dart';
+import 'package:chocobread/page/event/event_list.dart';
 import 'package:chocobread/page/form.dart';
 import 'package:chocobread/page/login.dart';
 import 'package:chocobread/page/nicknameset.dart';
@@ -294,7 +294,7 @@ class _HomeState extends State<Home> {
               onPressed: () {
                 Navigator.push(context,
                     MaterialPageRoute(builder: (BuildContext context) {
-                  return const EventPage();
+                  return const EventList();
                 }));
               },
               icon: const FaIcon(
